Restructure methods to files
