+++ lib/includeExternalDocument.js
