+++ lib/testAssertion.js
